2021.09.19(일)
 * 댓글 등록 API 수정
 * 대댓글 등록 API 추가

+++ src/main/java/kr/co/platform/util/auth/JwtAuthProvider.java
@@ -75,7 +75,7 @@ public class JwtAuthProvider {
         // 토큰 기반으로 유저의 정보 파싱
         Claims claims = Jwts.parser().setSigningKey(signatureKey).parseClaimsJws(token).getBody();
 
-        Long id = claims.get("id", Long.class);
+        long id = claims.get("id", Integer.class);
         String email = claims.get("email", String.class);
         String nickname = claims.get("nickname", String.class);
